Create MRIP species recreational effort indicator

This function creates an indicator with the proportion trips targeting a species of interest For new data queries, use MRIP Query Tool (https://www.fisheries.noaa.gov/data-tools/recreational-fisheries-statistics-queries) Query 'Time Series' under 'Effort Data'. Choose years of interest, summarize by Annual, Calendar Year, Atlantic Coast by State, all modes and areas Download csv as output

Usage

create_prop_sp_trips(
  total_trips,
  species_trips,
  states = c("MAINE", "CONNECTICUT", "MASSACHUSETTS", "NEW HAMPSHIRE", "NEW JERSEY",
    "NEW YORK", "RHODE ISLAND", "MARYLAND", "DELAWARE", "NORTH CAROLINA"),
  groupby_state = FALSE,
  return = TRUE
)

Arguments

total_trips

The MRIP trip data (R object mrip_effort) of total annual angler trips from downloaded csv file

species_trips

The directed trips data for species of interest (R object rec_trips) from function 'create_rec_trips'

states

States in which to filter data, from MRIP query 'ATLANTIC COAST BY STATE'

Value

Saves the R data object prop_sp_trips