Format data for upload to AKFIN web service

This function formats indicator data for upload to AKFIN web service

Usage

format_indicator(
  out_dir,
  submission_year,
  indicator_name,
  description,
  status,
  factors,
  implications,
  references,
  years,
  region,
  indicator_value,
  ...
)

Arguments

out_dir

dictory where the .txt file will be saved

submission_year

Current year of contribution submission. This is a integer value.

indicator_name

Name of the indicator. This is a primary key and must match what has been used in previous years. This is a text string.

description

Brief description of the indicator. Please make sure this description includes information on the spatial distribution of the indicator and how the data for the indicator are collected. The spatial resolution can be a cell size dimension (e.g., 5 km resolution for gridded data) or area of sampling for a survey. The data collection method can be brief (e.g., survey name and gear type, satellite sensor and version, stock assessment model output, fishery observer data, community reports, etc.) and can include a reference to methods detailed elswhere. This is a text string.

status

Information on the current status of the indicator in the context of historical trends. This is a text string.

factors

Information on the potential causes for observed trends and current status (focus on most recent year). This is a text string.

implications

Information that briefly answers these questions: What are the implications or impacts of the observed trends on the ecosystem or ecosystem components? What do the trends mean? Why are they important? How can this information be used to inform management decisions? This is a text string.

references

Include any full references that are associated with the indicator. This may include data references such as from an ERDDAP webpage or literature cited (plus the DOI where possible). This is a text string.

years

Years for the indicator contribution. This is a numeric vector.

region

List of spatial or temporal scales applicable to indicator. This is a character value.

indicator_value

Data values for the indicator contribution (must match the YEAR list length). This is a numeric vector.

...

not used

Value

A .txt file saved in out_dir