A wrapper function to summarise survdat data

This function is a wrapper for get_var_data, data_summary, and data_summary_5yr.

Usage

generate_table(x, variable, cap = "", type = "html")

Arguments

x

A survdat data frame or tibble, containing data on one species.

variable

The survdat measurement of interest. Must be a column name of the survdat data set.

cap

A caption for the table. Defaults to "".

type

The file type of the output. One of c("html", "word")

Value

A DT::datatable or a knitr::kable