Survdat data pull

A data pull of bottom trawl survey data from 1983-2024 using the survdat package.

Usage

survdat_subset

Format

A data frame with 16527 rows and 28 variables:

CRUISE6

integer Code uniquely identifying cruise. The first four digits indicate the year and the last two digit uniquely identify the cruise within the year.

STATION

integer Unique sequential order in which stations have been completed.

STRATUM

integer A predefined area where a net dredge, or other piece of gear was deployed. Code consists of 2 parts: Stratum group code number (2 bytes) and stratum number (3 bytes).

SVSPP

integer A standard code which represents a species caught in a trawl or dredge. Refer to the SVDBS.SVSPECIES_LIST

CATCHSEX

integer Code used to identify species that are sexed at the catch level.

LENGTH

integer Measured length of species in centimeters (cm).

TOW

integer Sequential number representing order in which station was selected within a stratum.

SVVESSEL

character Standard two character code for a survey vessel.

YEAR

integer Year in which cruise was conducted.

SEASON

character Season of the year in which cruise was conducted.

LAT

double Beginning latitude of tow in decimal degrees.

LON

double Beginning longitude of tow in decimal degrees.

EST_TOWDATE

character Date and time represented by Eastern Standard Time (EST) for the start of a tow or deployment.

DEPTH

integer A four digit number recording the average depth, to the nearest meter, during a survey gear deployment.

SURFTEMP

double Surface temperature of water (degrees Celsius).

SURFSALIN

double Salinity at water surface in practical salinity units (PSU).

BOTTEMP

double Bottom temperature (degrees Celsius).

BOTSALIN

double Bottom salinity in Practical Salinity Units (PSU).

ABUNDANCE

integer Expanded number of individuals of a species caught at a given station.

BIOMASS

double Expanded catch weight of a species caught at a given station.

NUMLEN

integer Expanded number of specimens at a given length.

INDID

integer A unique identifier for each fish sampled.

INDWT

double Individual weight (KG) of species being sampled.

SEX

integer Code indicating sex of fish or invertebrate species.

MATURITY

character Stage of maturation of the fish being sampled.

AGE

integer Age of specimen in years.

STOM_VOLUME

double Volume of the stomach contents of the fish sampled, measured to the nearest tenth of a cubic centimeter (cc).

STOM_WGT

double Stomach weight of an individual fish in grams.