Create species condition indicator

This function calculates species condition from a survdat data pull. Methods derived from Laurel Smith (https://github.com/Laurels1/Condition/blob/master/R/RelConditionEPU.R)

Usage

species_condition(data, LWparams, species.codes)

Arguments

data

data table. NEFSC survey data generated by 'survdat::get_survdat_data(channel, getBio = T, getLengths = T)'

LWparams

L-W parameters from Wigley et al. 2003 from 'data-raw/tech_memo_parameters_table_format.csv'

species.codes

data table. Species codes and names from 'data-raw/bottomtrawl_species_codes_names.csv'

Value

Returns R object condition, a data frame of species condition by EPU with scale.